test(reactive-rpc): 💍 move Reactive RPC E2E tests under the /src/reactive-rpc folder
