+++ package.json
@@ -53,8 +53,8 @@
     "test:cli:pointer": "./bin/json-pointer-test.js ./bin/json-pointer.js",
     "test:cli:patch": "./bin/json-patch-test.js ./bin/json-patch.js",
     "test:cli:pack": "./bin/json-pack-test.js ./bin/json-pack.js",
-    "test:reactive-rpc": "node -r ts-node/register/transpile-only src/__tests__/reactive-rpc/run.ts",
-    "test:reactive-rpc:jest": "TEST_E2E=1 jest --maxWorkers 1 --no-cache src/__tests__/reactive-rpc/",
+    "test:reactive-rpc": "node -r ts-node/register/transpile-only src/reactive-rpc/__tests__/e2e/run.ts",
+    "test:reactive-rpc:jest": "TEST_E2E=1 jest --maxWorkers 1 --no-cache src/reactive-rpc/__tests__/e2e/",
     "demo:json-patch": "ts-node src/json-patch/__demos__/json-patch.ts",
     "demo:json-pointer": "ts-node src/json-pointer/__demos__/json-pointer.ts",
     "demo:reactive-rpc:server": "ts-node src/reactive-rpc/__demos__/server.ts",
