+++ src/reactive-rpc/__tests__/e2e/uws/http/FetchRpcClient.spec.ts
@@ -2,16 +2,14 @@
  * @jest-environment node
  */
 
-import {ApiTestSetup, runApiTests} from '../../../../reactive-rpc/common/rpc/__tests__/runApiTests';
-import WebSocket from 'ws';
-import {RpcCodecs} from '../../../../reactive-rpc/common/codec/RpcCodecs';
-import {RpcMessageCodecs} from '../../../../reactive-rpc/common/codec/RpcMessageCodecs';
-import {Writer} from '../../../../util/buffers/Writer';
-import {Codecs} from '../../../../json-pack/codecs/Codecs';
-import {RpcMessageCodec} from '../../../../reactive-rpc/common/codec/types';
-import {JsonValueCodec} from '../../../../json-pack/codecs/types';
-import {RpcCodec} from '../../../../reactive-rpc/common/codec/RpcCodec';
-import {RpcPersistentClient, WebSocketChannel} from '../../../../reactive-rpc/common';
+import {ApiTestSetup, runApiTests} from '../../../../common/rpc/__tests__/runApiTests';
+import {RpcCodecs} from '../../../../common/codec/RpcCodecs';
+import {RpcMessageCodecs} from '../../../../common/codec/RpcMessageCodecs';
+import {Writer} from '../../../../../util/buffers/Writer';
+import {Codecs} from '../../../../../json-pack/codecs/Codecs';
+import {RpcMessageCodec} from '../../../../common/codec/types';
+import {JsonValueCodec} from '../../../../../json-pack/codecs/types';
+import {FetchRpcClient} from '../../../../common/rpc/client/FetchRpcClient';
 
 if (process.env.TEST_E2E) {
   const codecs = new RpcCodecs(new Codecs(new Writer()), new RpcMessageCodecs());
@@ -52,22 +50,17 @@ if (process.env.TEST_E2E) {
   for (const [protocolSpecifier, msgCodec, reqCodec, resCodec] of cases) {
     const setup: ApiTestSetup = async () => {
       const port = +(process.env.PORT || 9999);
-      const url = `ws://localhost:${port}/rpc`;
-      const codec = new RpcCodec(msgCodec, reqCodec, resCodec);
-      const client = new RpcPersistentClient({
-        codec,
-        channel: {
-          newChannel: () =>
-            new WebSocketChannel({
-              newSocket: () => new WebSocket(url, [codec.specifier()]) as any,
-            }),
-        },
+      const url = `http://localhost:${port}/rpc`;
+      const client = new FetchRpcClient({
+        url,
+        msgCodec,
+        reqCodec,
+        resCodec,
       });
-      client.start();
       return {client};
     };
-    describe(`protocol: application/x.${protocolSpecifier}`, () => {
-      runApiTests(setup);
+    describe(`Content-Type: application/x.${protocolSpecifier}`, () => {
+      runApiTests(setup, {staticOnly: true});
     });
   }
 } else {
