+++ src/reactive-rpc/__tests__/e2e/uws/ws/RpcPersistentClient.spec.ts
@@ -2,14 +2,16 @@
  * @jest-environment node
  */
 
-import {ApiTestSetup, runApiTests} from '../../../../reactive-rpc/common/rpc/__tests__/runApiTests';
-import {StreamingRpcClient} from '../../../../reactive-rpc/common';
-import {RpcCodecs} from '../../../../reactive-rpc/common/codec/RpcCodecs';
-import {RpcMessageCodecs} from '../../../../reactive-rpc/common/codec/RpcMessageCodecs';
-import {Writer} from '../../../../util/buffers/Writer';
-import {Codecs} from '../../../../json-pack/codecs/Codecs';
-import {RpcMessageCodec} from '../../../../reactive-rpc/common/codec/types';
-import {JsonValueCodec} from '../../../../json-pack/codecs/types';
+import {ApiTestSetup, runApiTests} from '../../../../common/rpc/__tests__/runApiTests';
+import WebSocket from 'ws';
+import {RpcCodecs} from '../../../../common/codec/RpcCodecs';
+import {RpcMessageCodecs} from '../../../../common/codec/RpcMessageCodecs';
+import {Writer} from '../../../../../util/buffers/Writer';
+import {Codecs} from '../../../../../json-pack/codecs/Codecs';
+import {RpcMessageCodec} from '../../../../common/codec/types';
+import {JsonValueCodec} from '../../../../../json-pack/codecs/types';
+import {RpcCodec} from '../../../../common/codec/RpcCodec';
+import {RpcPersistentClient, WebSocketChannel} from '../../../../common';
 
 if (process.env.TEST_E2E) {
   const codecs = new RpcCodecs(new Codecs(new Writer()), new RpcMessageCodecs());
@@ -48,37 +50,24 @@ if (process.env.TEST_E2E) {
   ];
 
   for (const [protocolSpecifier, msgCodec, reqCodec, resCodec] of cases) {
-    const contentType = 'application/x.' + protocolSpecifier;
     const setup: ApiTestSetup = async () => {
-      const client = new StreamingRpcClient({
-        send: async (messages) => {
-          const port = +(process.env.PORT || 9999);
-          const url = `http://localhost:${port}/rpc`;
-          reqCodec.encoder.writer.reset();
-          msgCodec.encodeBatch(reqCodec, messages);
-          const body = reqCodec.encoder.writer.flush();
-          try {
-            const response = await fetch(url, {
-              method: 'POST',
-              headers: {
-                'Content-Type': contentType,
-              },
-              body,
-            });
-            const buffer = await response.arrayBuffer();
-            const data = new Uint8Array(buffer);
-            const responseMessages = msgCodec.decodeBatch(resCodec, data);
-            client.onMessages(responseMessages as any);
-          } catch (err) {
-            // tslint:disable-next-line:no-console
-            console.error(err);
-          }
+      const port = +(process.env.PORT || 9999);
+      const url = `ws://localhost:${port}/rpc`;
+      const codec = new RpcCodec(msgCodec, reqCodec, resCodec);
+      const client = new RpcPersistentClient({
+        codec,
+        channel: {
+          newChannel: () =>
+            new WebSocketChannel({
+              newSocket: () => new WebSocket(url, [codec.specifier()]) as any,
+            }),
         },
       });
+      client.start();
       return {client};
     };
-    describe(`Content-Type: ${contentType}`, () => {
-      runApiTests(setup, {staticOnly: true});
+    describe(`protocol: application/x.${protocolSpecifier}`, () => {
+      runApiTests(setup);
     });
   }
 } else {
